Improved auto-refresh after watering
Percentages will now update correctly after the garden is watered either with a left click on the plant or the wold way. It will check to see if any wildlife was spotted, if it was then it will use the previous recorded last water time (as until it is watered again the plates which were not visited have not yet been checked).