Multiple bug fixes
Previous fix to the &amp; auto writer was implemented incorrectly and caused all wildlife that triggered that to prepend "xundefined" as they did not have multiple keys. This fixes that.

Also fixed an error caused by an if condition checking the wrong thing for how long wildlife stayed for. It was previously looking at something like "stayed for 8 minutes but" and was looking for what index "but" was at, and if it was not at a certain index it would assume that seconds followed minutes. This didn't actually impact anything as I had already caught his exception and returned 0, but it's fixed regardless.