Fixed error in getname
I had a typo and was checking the wrong index for two spaces, returning true despite there only being one space. Upon fix I realized I also needed to check for the arrow indactors in the name string in the event the user is in their own garden (the move left arrow).

The result of this fix is that certain foods will now give the correct "activetime".