0.8.3 - Notification bug fix
Previously the about tab only checked if it needed to remove the (1) notification if it had already been seen on this device, the problem happened when multiple devices happened and one saw it it wouldn't check if it needed to remove it on the others. It will now remove it each time the mushroom is clicked.